requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langchain-openai==0.1.23
langchain-pinecone==0.1.3
pinecone-client==5.0.1
openai==1.43.0
pydantic==2.8.2
tiktoken==0.7.0
python-multipart==0.0.9 